+++ ui/src/components/dashboard/Dashboard.tsx
@@ -1,8 +1,9 @@
 import { useEffect, useMemo, useReducer, useRef, useState, useSyncExternalStore } from 'react'
 import {
+  CircularProgress,
   FormControl,
   IconButton,
-  InputAdornment,
+  InputAdornment, Link,
   OutlinedInput,
   Stack,
   Tooltip,
@@ -19,12 +20,19 @@ import useConfig from '../../hooks/useConfig'
 import { Session } from '../../types/session'
 import { SessionStore } from '../../stores/sessionStore'
 import ExampleContainerButton, { ExampleContainerImageTag } from './ExampleContainerButton'
+import {filesize} from "filesize"
+import DockerHubApi from '../../libs/dockerHub/DockerHubApi'
 
 
 interface DashboardProps {
   ddUIToast: Toast
   connect: (session: Session)=>Promise<void>
   sessionStore: SessionStore
+  openUrl: (url: string)=>void
+}
+
+interface InfoTextImageSizeProps extends TypographyProps {
+  image: string
 }
 
 const UbuntuVNCDockerSessionName = 'example vnc container'
@@ -47,7 +55,40 @@ function InfoText(props: TypographyProps) {
 }
 
 
-export default function Dashboard({ ddUIToast, connect, sessionStore }: DashboardProps) {
+function InfoTextImageSize(props: InfoTextImageSizeProps) {
+  const dockerHubApi = new DockerHubApi()
+  const [imageSize, setImageSize] = useState<string | undefined>()
+  const { image: imageWithTag } = props
+
+  useEffect(() => {
+    setImageSize(undefined)
+
+    if (imageWithTag === '') {
+      setImageSize('ERROR')
+
+      return
+    }
+
+    const [image, tag] = imageWithTag.split(':')
+    const dockerRepo = dockerHubApi.repository(image)
+    dockerRepo.getTag(tag)
+              .then(imageInfo => setImageSize(filesize(imageInfo.full_size, {standard: 'jedec'})))
+  }, [imageWithTag])
+
+  return <InfoText>Image size = {
+    imageSize === undefined
+      ? <CircularProgress size={20} sx={{
+          display: 'inline-block',
+          maxWidth: '50px',
+          width: '50px',
+          verticalAlign: 'middle',
+        }} />
+      : <Typography component="span" sx={{display: 'inline-block', textDecoration: 'underline'}}>{ imageSize }</Typography>
+  }</InfoText>
+}
+
+
+export default function Dashboard({ ddUIToast, openUrl, connect, sessionStore }: DashboardProps) {
   const [loading, setLoading] = useState<boolean>(true)
   const sessions = useSyncExternalStore(sessionStore.subscribe, sessionStore.getSnapshot)
   const exampleRunInputRef = useRef<HTMLInputElement>(null)
@@ -290,6 +331,14 @@ export default function Dashboard({ ddUIToast, connect, sessionStore }: Dashboar
           <InfoText>VNC connect Password = { proxyContainerPassword }</InfoText>
           <InfoText>Docker Container Name = ubuntu_vnc</InfoText>
           <InfoText>VNC Port = 5901</InfoText>
+          <InfoTextImageSize image={ubuntuVNCDockerImage} />
+          <InfoText>
+            GitHub ={" "}
+            <Link
+              component="button"
+              onClick={() => openUrl(`https://github.com/pgmystery/docker-extension-vnc/tree/main/docker/vnc_ubuntu/${exampleContainerTag}`)}
+            >{ubuntuVNCDockerImage}</Link>
+          </InfoText>
         </FormControl>
 
         <ExampleContainerButton