+++ ui/src/hooks/useBackdrops/useBackdrop.tsx
@@ -1,32 +1,80 @@
-import React, { useContext, useMemo } from 'react'
+import React, { useContext, useEffect, useMemo, useState } from 'react'
 import { BackdropContext } from './BackdropContext'
 import { BackdropProps } from '@mui/material/Backdrop/Backdrop'
 import { Backdrop, CircularProgress } from '@mui/material'
+import * as crypto from 'node:crypto'
 
 
-export type BackdropComponentProps = Omit<BackdropProps, 'open'>
-export type CreateUseBackdropHook = (backdropProps?: BackdropComponentProps)=>UseBackdropHook
+export type BackdropComponentBackdropProps = Omit<BackdropProps, 'open'>
+
 export interface UseBackdropHook {
-  showBackdrop: ShowBackdrop
-  isBackdropShowing: boolean
+  open: ShowBackdrop
+  isOpen: boolean
 }
-export type ShowBackdrop = <T>(asyncCallback: () => Promise<T>) => Promise<T>
+
+export type ShowBackdrop = <T>(asyncCallback: () => Promise<T>, currentBackdropProps?: BackdropComponentBackdropProps) => Promise<T>
+
+interface Backdrop {
+  id: string
+  props: BackdropComponentBackdropProps
+}
 
 
-export default function useBackdrop(backdropProps?: BackdropComponentProps): UseBackdropHook {
-  const backdropContext = useContext(BackdropContext)
+export default function useBackdrop(backdropProps: BackdropComponentBackdropProps = {}): UseBackdropHook {
+  const id = useMemo(() => crypto.randomBytes(20).toString('hex'), [])
+  const backdropHandlerDispatch = useContext(BackdropContext)
 
-  if (backdropContext === null || backdropContext === undefined) {
+  if (backdropHandlerDispatch === null || backdropHandlerDispatch === undefined) {
     throw new Error('context "BackdropContext" was used without a Provider')
   }
 
-  return useMemo(() => backdropContext(backdropProps), [backdropProps])
+  const [backdrops, setBackdrops] = useState<Backdrop[]>([])
+
+  useEffect(() => {
+    return () => {
+      backdropHandlerDispatch({
+        type: 'UNMOUNT',
+        id,
+      })
+    }
+  }, [])
+
+  useEffect(() => {
+    backdropHandlerDispatch({
+      type: 'SET',
+      id,
+      backdrops: backdrops.map(backdrop => backdrop.props),
+    })
+  }, [backdrops])
+
+  return useMemo(() => {
+    return {
+      open: function<T> (asyncCallback: ()=>Promise<T>, currentBackdropProps: BackdropComponentBackdropProps = {}) {
+        const backdropId = crypto.randomBytes(20).toString('hex')
+        const backdrop: Backdrop = {
+          id: backdropId,
+          props: {
+            ...backdropProps,
+            ...currentBackdropProps,
+          }
+        }
+
+        setBackdrops([
+          ...backdrops,
+          backdrop,
+        ])
+
+        return asyncCallback().finally(() => setBackdrops(backdrops.filter(backdrop => backdrop.id !== id)))
+      },
+      isOpen: backdrops.length > 0
+    }
+  }, [backdrops])
 }
 
 
-export function UseBackdropComponent({ backdropProps }: { backdropProps?: BackdropComponentProps }) {
+export function UseBackdropComponent(props: BackdropComponentBackdropProps) {
   return (
-    <Backdrop { ...backdropProps } open>
+    <Backdrop { ...props } open>
       <CircularProgress />
     </Backdrop>
   )