+++ ui/src/libs/dockerHub/DockerHubApi.ts
@@ -1,0 +1,8 @@
+import Repository from './dockerHubApi/Repository'
+
+
+export default class DockerHubApi {
+  repository(repository: string) {
+    return new Repository(repository)
+  }
+}