+++ CHANGELOG.md
@@ -2,6 +2,14 @@
 
 History of all released versions and changes
 
+## v1.3.0
+- Show the Image size from the selected Example Docker Image on Dashboard.
+- Add a screenshot-button for making a screenshot from the canvas and save it on the user's computer.
+- Add a record-button to record the canvas as a video and save it on the user's computer.
+- Disable some session buttons if there is no session (disconnected).
+- NoVNC-Proxy: Update Websockify version from v0.12.0 -> v0.13.0.
+- Add GitHub-url on Dashboard for the selected example Docker-Image.
+
 ## v1.2.0
 - Fix bug with "Toggle keys" under a windows machine.
 - Add a new button for copy the url of the proxy-websocket.